requirements.txt
gradio
langchain
pypdf
sentence-transformers
faiss-cpu
openai
python-dotenv